Create README.md
The goal of the project was to help any banking organization find a simple and effective predictive model that would allow them to determine whether or not their customers could make timely credit card payments. With the invention of the credit card, banks were more focused on the number of customers who used their credit service, but the drawback of them not being able to pay back the credit on time quickly became an issue. A system was needed to effectively decide the credit limit to be allowed to a person based on his previous credit history, concentrating on this issue we were motivated to consider various parameters such as sex, age , level of education along with customer details and the credit history of the past 6 months to give a reliable and effective prediction if a customer is able to pay the credit for the next (7th) month.

algorithm
built using three models:
• Neural Network
• Decision Tree
• K-Nearest Neighbors
• Naïve Bayes